WIP: Add the initial templating
This adds the initial version of the templating including a load
if CSS and Javascript that needs to be deleted once I work out
which page actually needs which scripts/css. At the moment every page
will include all the CSS and javascript which isn't ideal.

I also added flask-menu to the project and used it to render
the nav bar with an example of a nested page config.

Now I should be able to deconstruct the base template to create the
login flow. The current index.html makes no sense but it is just
a nice example of how to build the sections.