Add poor mans class for tracking RSVP state changes
I'm not totally happy with this implementation but it does seem
pragmatically like the simplest thing to do. I explicitly store
the set of changes in a new table. The really ugly thing is I had
to create an extra frozen RSVP class to store the RSVP state because
sqlalchemy is annoying and changes the ORM state silently when I
update the state. So I can;t just hold a reference to the "before"
state because sqlalchemy will expire the state and refetch it when
I go to commit. I tried various tricks around expire_on_commit
and new transactions etc to stop this (including explicitly
expunging the ORM objects) but none of it worked. So now I have
a frozen object class that feels awfully like it should be a database
table. BUT I don't like the lack of referential integrity around
storing a list of attending guest IDs...