Add relationship helper attributes to RSVP changes for guest objects
Had to use some fancy sqlalchemy features to get this done.
Basically do a subquery join on the unnested array values which
is pretty inefficient but this table should be pretty small. Wouldn't
work if it was really millions of entries but it will be in the
order of 1000 max

Note that accessing these attributes is a bit expensive as
I cannot get them to pre-load at the moment